Moves contract-client-gen code to its own file
